// src/app/page/master/zodiak/zodiak.interface.ts
    import {EventEmitter} from '@angular/core';

    export interface Zodiak {
	kdPangkat?: any;
	namaPangkat?: string;
	reportDisplay?: string;
	noUrut?: number;
	kdGolonganPegawai?: number;
	ruang?: string;
	kdPangkatHead?: number;
	kodeExternal?: string;
	namaExternal?: string;
	kdDepartemen?: string;
	statusEnabled?: boolean;
	noRec?: string;
	label?: any;
  	value?: any;
  	namaGolonganPegawai?: string;
  	id?:any;
  	kode?:any;
  	version?:any;
    }